isolated settings modals into separate files and cleaned design
